# docs/source/developer_documentation/retriever_api_specs.yaml
openapi: 3.1.1
info:
  title: mmore Retriever API
  description: |
    This API is based on the OpenAPI 3.1 specification. You can find out more about Swagger at [https://swagger.io](https://swagger.io).

    ## Overview

    This API defines the retriever API of mmore, handling:

    1. **File Operations** - Direct file management within mmore.
    2. **Context Retrieval** - Semantic search based on the subset of documents that the user wants.

    ## API Versioning

    All requests must be prefixed with `/v1` for this current version of the API.

    ## Roadmap & Considerations
    - Authorization layer for mmore operations
    - Permission control on some operation such as file deletion
    - Advanced query for retrieval (send more sophisticated object, for instance representing a whole conversation, rather than one string)

  version: 1.0.0
servers:
  - url: /v1
    description: API server
paths:
  /files:
    post:
      tags:
        - File Operations
      summary: Upload a file
      description: |
        Upload a new file

         **Requirements**:
        - Unique fileId
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                fileId:
                  type: string
                  description: Unique identifier for the file
                file:
                  type: string
                  format: binary
                  description: The file content
                collectionName:
                  type: string
                  description: |
                    Collection to index into. Defaults to the configured
                    `collection_name`. Created if it does not exist.
              required:
                - fileId
                - file
      responses:
        "201":
          description: File successfully uploaded
  /files/bulk:
    post:
      tags:
        - File Operations
      summary: Upload multiple files
      description: |
        Upload several files at once, one independent job per file. All files
        are indexed into the same collection.

         **Requirements**:
        - One unique ID per file, in the same order as the files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                listIds:
                  type: array
                  items:
                    type: string
                  description: |
                    IDs of the files, in the same order as `files`. Each
                    entry may also hold several comma-separated IDs.
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: The files' contents
                collectionName:
                  type: string
                  description: |
                    Collection to index into. Defaults to the configured
                    `collection_name`. Created if it does not exist.
              required:
                - listIds
                - files
      responses:
        "202":
          description: Per-file outcome, either a jobId or an error
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      type: object
                      properties:
                        fileId:
                          type: string
                        jobId:
                          type: string
                        error:
                          type: string
        "400":
          description: Number of IDs does not match number of files
...
  /files/{fileId}:
    put:
      tags:
        - File Operations
      summary: Update a file
      description: Replace an existing file with a new version
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The new file content
                collectionName:
                  type: string
                  description: |
                    Collection the file is indexed in. Defaults to the
                    configured `collection_name`.
              required:
                - file
      responses:
        "200":
          description: File successfully updated
        "404":
          description: File or collection not found

    delete:
      tags:
        - File Operations
      summary: Remove a file
      description: |
        Delete a file from the system.

        **Warning**: Not limited operation.
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: query
          required: false
          description: |
            Collection the file is indexed in. Defaults to the configured
            `collection_name`.
          schema:
            type: string
      responses:
        "200":
          description: File successfully deleted
        "404":
          description: File or collection not found

    get:
      tags:
        - File Operations
      summary: Download a file
      description: Download a file from the system
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
        - name: collectionName
          in: query
          required: false
          description: |
            Collection the file is indexed in. Defaults to the configured
            `collection_name`.
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File or collection not found
...